+++ setup/telescopes.py
@@ -1614,7 +1614,7 @@ frost_param = {
                                          # pp_prepare
     'object'               : 'OBJECT',  # object name keyword 
     'filter'               : 'FILTER',  # filter keyword
-    'filter_translations'  : {'clear': 'R'},
+    'filter_translations'  : {'clear': 'V'},
                              # filtername translation dictionary
     'exptime'              : 'EXPTIME', # exposure time keyword (s)
     'airmass'              : 'AIRMASS', # airmass keyword