+++ doc/changelog.rst
@@ -3,6 +3,8 @@ Changelog
 
 Major changes to the pipeline since 2016-10-01 (see `Mommert 2017`_) are documented here.
 
+* 2017-06-01: extraction of serendipitously observed targets (asteroids and variable stars) implemented in ``pp_distill`` 
+
 * 2017-03-19: if no filtername is provided (``None``) or the ``-instrumental`` option of ``pp_calibrate`` is used, ``pp_run`` will complete all pipeline tasks using these instrumental magnitudes 
 
 * 2017-03-16: implementation of `Pan-STARRS DR1`_ for the photometric calibration; currently, PP uses a home-built access of MAST at STScI, which is limited to catalog queries with a maximum cone radius of 0.5 deg; please note that this kind of query is rather slow compared to Vizier queries of SDSS or APASS